refactor moxy adding method
