+++ app/src/main/java/com/itransition/trainingprojectandroid/HomeActivity.kt
@@ -2,8 +2,7 @@ package com.itransition.trainingprojectandroid
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import com.itransition.trainingprojectandroid.mvp.ListFragment
+import com.itransition.trainingprojectandroid.mvp.MvpFragment
 
 class HomeActivity : AppCompatActivity() {
 
@@ -16,7 +15,7 @@ class HomeActivity : AppCompatActivity() {
     private fun setFragmentContainer() {
         supportFragmentManager
             .beginTransaction()
-            .replace(R.id.fragmentContainer, ListFragment() as Fragment)
+            .replace(R.id.fragmentContainer, MvpFragment())
             .addToBackStack(null)
             .commit()
     }
