+++ app/src/main/java/com/itransition/trainingprojectandroid/mvp/MvpFragment.kt
@@ -5,17 +5,20 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.itransition.trainingprojectandroid.mvp.recyclerview.ListRecyclerModel
-import com.itransition.trainingprojectandroid.mvp.recyclerview.RecyclerListAdapter
 import com.itransition.trainingprojectandroid.R
+import com.itransition.trainingprojectandroid.mvp.interfaces.SetDataInterfaces
+import com.itransition.trainingprojectandroid.mvp.presenter.SetDataPresenter
+import com.itransition.trainingprojectandroid.mvp.recyclerview.RecyclerListAdapter
+import com.itransition.trainingprojectandroid.mvp.recyclerview.RecyclerModel
 import kotlinx.android.synthetic.main.fragment_list.*
-import moxy.MvpFragment
+import moxy.MvpAppCompatFragment
 import moxy.presenter.InjectPresenter
 
-class ListFragment : MvpFragment(), ListInterfaces {
+class MvpFragment : MvpAppCompatFragment(),
+    SetDataInterfaces {
 
     @InjectPresenter
-    lateinit var presenter: ListPresenter
+    lateinit var setDataPresenter: SetDataPresenter
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
@@ -23,13 +26,10 @@ class ListFragment : MvpFragment(), ListInterfaces {
         return inflater.inflate(R.layout.fragment_list, container, false)
     }
 
-    override fun getData(value: ArrayList<ListRecyclerModel>) {
+    override fun addValue(value: ArrayList<RecyclerModel>) {
         recycler_view.layoutManager = LinearLayoutManager(this.context)
         val recyclerListAdapter = context?.let { RecyclerListAdapter(value, it) }
         recycler_view.adapter = recyclerListAdapter
     }
 
-    override fun addValue(value: ArrayList<ListRecyclerModel>) {
-        TODO("Not yet implemented")
-    }
 }
