+++ app/src/main/java/com/itransition/trainingprojectandroid/mvp/recyclerview/RecyclerListAdapter.kt
@@ -5,17 +5,16 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.itransition.trainingprojectandroid.R
-import com.itransition.trainingprojectandroid.listfragment.recyclerview.ListRecyclerModel
-import com.itransition.trainingprojectandroid.listfragment.recyclerview.RecyclerListViewHolder
 
 class RecyclerListAdapter(
-    var value: ArrayList<ListRecyclerModel>,
+    var value: ArrayList<RecyclerModel>,
     var context: Context
 ) : RecyclerView.Adapter<RecyclerListViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerListViewHolder {
         return RecyclerListViewHolder(
-            LayoutInflater.from(context).inflate(R.layout.list_recycler_view, parent, false))
+            LayoutInflater.from(context).inflate(R.layout.list_recycler_view, parent, false)
+        )
     }
 
     override fun getItemCount(): Int {
