+++ app/src/main/java/com/itransition/trainingprojectandroid/mvp/recyclerview/RecyclerListViewHolder.kt
@@ -2,12 +2,11 @@ package com.itransition.trainingprojectandroid.mvp.recyclerview
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import com.itransition.trainingprojectandroid.listfragment.recyclerview.ListRecyclerModel
 import com.itransition.trainingprojectandroid.setTextVisibility
 import kotlinx.android.synthetic.main.list_recycler_view.view.*
 
 class RecyclerListViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    fun bind(model: ListRecyclerModel) {
+    fun bind(model: RecyclerModel) {
         itemView.title.text = model.title.toString()
         itemView.description.text = model.description
     }
