+++ app/src/main/java/com/itransition/trainingprojectandroid/mvp/recyclerview/RecyclerModel.kt
@@ -1,0 +1,5 @@
+package com.itransition.trainingprojectandroid.mvp.recyclerview
+
+import com.itransition.trainingprojectandroid.mvp.interfaces.SetDataInterfaces
+
+class RecyclerModel(var title: Int, var description: String)
